# tree-map/src/app/api/tree-requests/route.ts
import { NextRequest, NextResponse } from 'next/server';

export async function POST(request: NextRequest) {
  try {
    const { address, numTrees, location, lat, lng } = await request.json();

    // TODO: Add validation
    if (!address || !numTrees) {
      return NextResponse.json(
        { error: 'Address and number of trees are required' },
        { status: 400 },
      );
    }

    const formattedAddress = prepareAddress(address);

    const response = await fetch(`${process.env.API_URL}/tree-requests`, {
      method: 'POST',
      headers: {
        'Content-Type': 'application/json',
      },
      body: JSON.stringify({ address: formattedAddress, numTrees, location, lat, lng }),
    });
    const data = await response.json();
    if (data.error) {
      return NextResponse.json({ error: data.error }, { status: 400 });
    }

    return NextResponse.json({
      address,
      numTrees,
      location,
      srNumber: data.srNumber,
      lat,
      lng,
    });
  } catch (error) {
    console.error(error);
    return NextResponse.json(
      { error: error instanceof Error ? error.message : 'Unknown error' },
      { status: 500 },
    );
  }
}

function prepareAddress(address: string): string {
  if (
    !address.endsWith('Chicago, IL') &&
    !address.endsWith('Chicago, IL, USA') &&
    !address.endsWith('Chicago, Illinois, USA')
  ) {
    throw new Error('Address must be in Chicago');
  }
  const parts = address.split(',')[0].split(' ');
  const newAddress = {
    houseNumber: parts[0],
    streetDirection: parts[1].slice(0, 1),
    streetName: parts[2],
    streetType: streetTypeToAbbreviation(parts[3]),
  };
  return Object.values(newAddress).join(' ');
}

function streetTypeToAbbreviation(streetType: string): string {
  const streetTypeLower = streetType.toLowerCase();
  if (streetTypeLower === 'avenue' || streetTypeLower === 'ave') {
    return 'Ave';
  }
  if (streetTypeLower === 'street') {
    return 'St';
  }
  if (streetTypeLower === 'market') {
    return 'Mkt';
  }
  if (streetTypeLower === 'road') {
    return 'Rd';
  }
  if (streetTypeLower === 'drive') {
    return 'Dr';
  }
  if (streetTypeLower === 'place') {
    return 'Pl';
  }
  if (streetTypeLower === 'court') {
    return 'Ct';
  }
  if (streetTypeLower === 'parkway' || streetTypeLower === 'pkwy') {
    return 'Pkwy';
  }
  if (streetTypeLower === 'expressway' || streetTypeLower === 'expwy') {
    return 'Expwy';
  }
  if (streetTypeLower === 'highway' || streetTypeLower === 'hwy') {
    return 'Hwy';
  }
  if (streetTypeLower === 'circle' || streetTypeLower === 'cir') {
    return 'Cir';
  }
  if (streetTypeLower === 'terrace' || streetTypeLower === 'ter') {
    return 'Ter';
  }
  if (streetTypeLower === 'boulevard' || streetTypeLower === 'blvd') {
    return 'Blvd';
  }
  if (streetTypeLower === 'way' || streetTypeLower === 'wy') {
    return 'Way';
  }
  if (streetTypeLower === 'square' || streetTypeLower === 'sq') {
    return 'Sq';
  }
  return streetType.slice(0, 2);
}